CLI: refactor, move all bindings generation to its own module
